+++ Changelog.md
@@ -2,13 +2,14 @@
 - Version bump (25.09)
 - Build scripts now use python module mode
 - Platform binaries now only include necessary scripts
-- Swapped to glob method of checking & deleting cached ft-identifying-tablet-devices.html file
+- Swapped to glob method of checking & deleting cached ft-identify-tablet-devices.html file
 - Added line skip before "Updating Dependencies" section of update log
 - Fixed incorrect append command when adding `main.py` to modules
 - Fixed Linux/macOS versions crashing due to a Windows workaround #79 
 - Replaced os.startfile with a notepad command since it's not cross platform and causes incompatibility on Linux/macOS
 - Build scripts will now use the version file to automatically set the version variable
 - Added build directories to .gitignore
+- Updated Amazon device docs link for identifer scripts
 
 ## Updated main, build, version | Date: 07/21/25
 - Version bump (25.07.1)